requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
autopep8==2.0.1
black==23.1.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coverage==7.1.0
cryptography==38.0.4
daphne==4.0.0
dj-database-url==1.2.0
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.8.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
msgpack==1.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
zope.interface==5.5.2
